split lua config
