+++ scripts/symlinks.zsh
@@ -72,7 +72,7 @@ mkdir -p $user_dir/.config/nvim
 ln -sf $config_dir/nvim/init.vim $user_dir/.config/nvim/init.vim
 ln -sf $config_dir/nvim/config.vim $user_dir/.config/nvim/config.vim
 ln -sf $config_dir/nvim/config.lua $user_dir/.config/nvim/config.lua
-ln -sf $config_dir/nvim/coc-settings.json $user_dir/.config/nvim/coc-settings.json
+ln -sf $config_dir/nvim/lua $user_dir/.config/nvim/lua
 
 # picom
 mkdir -p $user_dir/.config/picom
